refactor: isolate PDF provenance conservation
